requirements.txt
echo "fastapi==0.78.0
httpx==0.23.0
pydantic==1.9.0
tortoise-orm==0.19.0
aerich==0.6.3
uvicorn==0.17.6
python-dotenv==0.20.0" > requirements.txt